requirements.txt
# ── BioAesthetic — Python Dependencies ────────────────────────────────────────

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9          # file uploads

# Validation
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0                  # async PostgreSQL driver
alembic==1.13.2                  # migrations

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# ML / CV
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
scikit-learn==1.5.1
pandas==2.2.2

# Phase 2 (CNN) — install separately if USE_ML_INFERENCE=True
# torch==2.4.0
# torchvision==0.19.0

# Caching
redis==5.0.8
aioredis==2.0.1

# AWS / S3
boto3==1.35.0
botocore==1.35.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2                    # async test client

# Utilities
python-dotenv==1.0.1
